driver-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
pydantic==2.4.2
grpcio==1.59.3
grpcio-tools==1.59.3
prometheus-client==0.21.1
